Add my assets

+++ example_create.sh
@@ -15,29 +15,44 @@
 # - mkdir -p htmldir && cd htmldir
 # - sh example_create.sh
 
-# path must be absolute.
-reposdir="/var/www/domains/git.codemadness.nl/home/src"
-curdir="$(pwd)"
+# path must be absolute
+reposdir="/srv/git"
+webdir="/srv/git/html"
+defaultdir="/usr/local/share/doc/stagit"
 
-# make index.
-stagit-index "${reposdir}/"*/ > "${curdir}/index.html"
+mkdir -p "$webdir" || exit 1
 
-# make files per repo.
-for dir in "${reposdir}/"*/; do
-	# strip .git suffix.
-	r=$(basename "${dir}")
-	d=$(basename "${dir}" ".git")
-	printf "%s... " "${d}"
+# set assets if not already there
+ln -s "$defaultdir/style.css" "$webdir/style.css" 2> /dev/null
+ln -s "$defaultdir/logo.png" "$webdir/logo.png" 2> /dev/null
+ln -s "$defaultdir/favicon.ico" "$webdir/favicon.ico" 2> /dev/null
 
-	mkdir -p "${curdir}/${d}"
-	cd "${curdir}/${d}" || continue
-	stagit -c ".cache" "${reposdir}/${r}"
-
-	# symlinks
-	ln -sf log.html index.html
-	ln -sf ../style.css style.css
-	ln -sf ../logo.png logo.png
-	ln -sf ../favicon.ico favicon.ico
-
-	echo "done"
+# clean
+for dir in "$webdir/"*/; do
+    rm -rf "$dir"
 done
+
+repos=""
+
+# make files per repo
+for dir in "$reposdir/"*.git/; do
+    [ ! -f "$dir/git-daemon-export-ok" ] && continue
+    repos="$repos $dir"
+
+    # strip .git suffix
+    r=$(basename "$dir")
+    d=$(basename "$dir" ".git")
+    printf "%s... " "$d"
+
+    mkdir -p "$webdir/$d"
+    cd "$webdir/$d" || continue
+    stagit -c ".stagit-build-cache" "$reposdir/$r"
+
+    # symlinks
+    ln -sf log.html index.html
+
+    echo "done"
+done
+
+# make index
+echo "$repos" | xargs stagit-index > "$webdir/index.html"
